+++ src/components/buttons/index.js
@@ -1,5 +1,5 @@
 import Btn from './Button'
-import BtnToggle from './ButtonToggle.vue'
+import BtnToggle from './ButtonToggle'
 import SpeedDial from './SpeedDial'
 
 export default {
